Create Generating Random Bits
Write a main program which will use flags -n -s -t and -h. The program can be called with one or more flags
Take N as input on command line with the "-n" flag
Take the seed as input with the "-s" flag
Dynamically set the seed to the current time using if the program is called with the "-t" flag
Compare the Hamming distance between two input bit stream files  when using the -h flag.